linkedin 1st round problem
There are ‘n’ ticket windows in the railway station, ith window has ai
tickets available. Price of a ticket is equal to the number of tickets
remaining in that window at that time. When ‘m’ tickets have been sold,
what’s the maximum amount of money the railway station can earn?